password-manager-main/password-manager-main/requirement.txt
Django==5.1.3
djangorestframework==3.15.2  => Working with API
pyotp==2.9.0 => working with OTP
cryptography==43.0.3    => Working with Incryption Password
django-cors-headers==4.6.0
